Separate integration/UI and unit tests
(cherry picked from commit b196ab8eeb65e8f688c6dda8ceb99eb7439acc2c)
