+++ GitExtensions.sln
@@ -209,6 +209,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".tx", ".tx", "{EFBCA0B9-9E0
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeleteUnusedBranchesTests", "UnitTests\Plugins\DeleteUnusedBranchesTests\DeleteUnusedBranchesTests.csproj", "{3FD5B353-17C0-4122-A693-27B707D06EFF}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "IntegrationTests", "IntegrationTests", "{8E1CE2CD-4099-43F1-94A9-0E4303D4623D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0E85F329-9297-4B9A-8CDD-CCE7601F1F32}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UITests", "IntegrationTests\UITests\UITests.csproj", "{27545CC5-7FFD-4F3F-8A33-8F3EF0CAC90C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -397,6 +403,10 @@ Global
 		{3FD5B353-17C0-4122-A693-27B707D06EFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3FD5B353-17C0-4122-A693-27B707D06EFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3FD5B353-17C0-4122-A693-27B707D06EFF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27545CC5-7FFD-4F3F-8A33-8F3EF0CAC90C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27545CC5-7FFD-4F3F-8A33-8F3EF0CAC90C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27545CC5-7FFD-4F3F-8A33-8F3EF0CAC90C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27545CC5-7FFD-4F3F-8A33-8F3EF0CAC90C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -429,6 +439,7 @@ Global
 		{F01BB9F8-40F5-4D19-A8C1-9ED8E30107B6} = {D3E0C080-5AC9-45DD-91A5-4933956C3560}
 		{1DC7D403-484B-43B4-B017-1356397A32CB} = {0E859634-B3DF-46AC-A73C-D49734F19D6A}
 		{39228D7F-9239-4A2D-8A6F-4D0F7DEB9178} = {F1B97F5D-CA42-4280-94A1-CE412BEA20DF}
+		{020F800B-9170-4BB3-83C9-FCF83596D83C} = {0E85F329-9297-4B9A-8CDD-CCE7601F1F32}
 		{E0BC2504-6BBB-4197-95EA-A2F7542BC15F} = {020F800B-9170-4BB3-83C9-FCF83596D83C}
 		{1CD494C1-0BBA-44C3-BDAD-EE241975B223} = {020F800B-9170-4BB3-83C9-FCF83596D83C}
 		{C2276A26-13DF-485A-8B2C-414CF507798F} = {020F800B-9170-4BB3-83C9-FCF83596D83C}
@@ -449,6 +460,8 @@ Global
 		{D017F26A-694F-4127-B917-9B1BD3F01E2F} = {472B72D9-9336-4C8B-A199-300F0CA7669E}
 		{EFBCA0B9-9E08-4464-BEAA-DAF2FF85C6AB} = {D017F26A-694F-4127-B917-9B1BD3F01E2F}
 		{3FD5B353-17C0-4122-A693-27B707D06EFF} = {C2276A26-13DF-485A-8B2C-414CF507798F}
+		{8E1CE2CD-4099-43F1-94A9-0E4303D4623D} = {0E85F329-9297-4B9A-8CDD-CCE7601F1F32}
+		{27545CC5-7FFD-4F3F-8A33-8F3EF0CAC90C} = {8E1CE2CD-4099-43F1-94A9-0E4303D4623D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {ECECB7EA-BCCF-44AD-B63E-C2FA45589FB0}
