+++ GitUI/Properties/AssemblyInfo.cs
@@ -4,5 +4,6 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyDescription("GitExtensions UI")]
 
 [assembly: InternalsVisibleTo("GitUITests")]
+[assembly: InternalsVisibleTo("UITests")]
 [assembly: InternalsVisibleTo("TranslationApp")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // required for NSubstitute for mocking internal members
