+++ IntegrationTests/UITests/CommandsDialogs/FormEditorTests.cs
@@ -3,12 +3,13 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using CommonTestUtils;
 using GitCommands;
 using GitUI;
 using GitUI.CommandsDialogs;
 using NUnit.Framework;
 
-namespace GitUITests.CommandsDialogs
+namespace GitExtensions.UITests.CommandsDialogs
 {
     [Apartment(ApartmentState.STA)]
     public class FormEditorTests
