+++ UnitTests/GitCommandsTests/RevisionReaderTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Text;
+using CommonTestUtils;
 using FluentAssertions;
 using GitCommands;
 using NUnit.Framework;
