+++ UnitTests/GitUITests/CommandsDialogs/FormOpenDirectoryTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.IO;
 using System.Threading;
+using CommonTestUtils;
 using FluentAssertions;
 using GitCommands.UserRepositoryHistory;
 using GitUI.CommandsDialogs.BrowseDialog;
