+++ UnitTests/GitUITests/GitUITests.csproj
@@ -64,22 +64,11 @@
     <Compile Include="Avatars\InitialsAvatarGeneratorTests.cs" />
     <Compile Include="Avatars\BackupAvatarProviderTests.cs" />
     <Compile Include="BranchTreePanel\GitRefMenuItemsTest.cs" />
-    <Compile Include="NBugReports\BugReportFormTests.cs" />
     <Compile Include="CancellationTokenSequenceTests.cs" />
     <Compile Include="CommandsDialogs\FormAddSubmoduleTests.cs" />
-    <Compile Include="CommandsDialogs\FormBrowseTests.LeftPanel.ReorderNodes.cs" />
-    <Compile Include="CommandsDialogs\FormBrowseTests.LeftPanel.Submodules.cs" />
-    <Compile Include="CommandsDialogs\FormBrowseTests.LeftPanel.Remotes.cs" />
-    <Compile Include="CommandsDialogs\FormBrowseTests.cs" />
     <Compile Include="CommandsDialogs\FormOpenDirectoryTests.cs" />
-    <Compile Include="CommandsDialogs\FormPullTests.cs" />
-    <Compile Include="CommandsDialogs\FormInitTests.cs" />
-    <Compile Include="CommandsDialogs\FormCommitTests.cs" />
-    <Compile Include="CommandsDialogs\FileStatusListTests.cs" />
-    <Compile Include="CommandsDialogs\FormEditorTests.cs" />
     <Compile Include="CommandsDialogs\FormFileHistoryControllerTests.cs" />
     <Compile Include="CommandsDialogs\PathFormatterTests.cs" />
-    <Compile Include="CommandsDialogs\ReferenceRepository.cs" />
     <Compile Include="CommandsDialogs\RevisionDiffContextMenuControllerTests.cs" />
     <Compile Include="CommandsDialogs\RevisionFileTreeControllerTests.cs" />
     <Compile Include="CommandsDialogs\RevisionDiffControllerTests.cs" />
@@ -99,13 +88,11 @@
     <Compile Include="ControlUtilTests.cs" />
     <Compile Include="LinqExtensionsTests.cs" />
     <Compile Include="Script\ScriptOptionsParserTests.cs" />
-    <Compile Include="Script\ScriptRunnerTests.cs" />
     <Compile Include="SpellChecker\WordAtCursorExtractorTests.cs" />
     <Compile Include="SplitterManagerTest.cs" />
     <Compile Include="StringBuilderExtensionsTests.cs" />
     <Compile Include="ThreadHelperTests.cs" />
     <Compile Include="TranslationTest.cs" />
-    <Compile Include="UITest.cs" />
     <Compile Include="UserControls\BlameControlTests.cs" />
     <Compile Include="UserControls\AuthorRevisionHighlightingTests.cs" />
     <Compile Include="UserControls\ConsoleEmulatorOutputControlFixture.cs" />
@@ -115,7 +102,6 @@
     <Compile Include="Hotkey\HotkeySettingsManagerFixture.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UserControls\RepoStateVisualiserTests.cs" />
-    <Compile Include="UserControls\RevisionGrid\CopyContextMenuItemTests.cs" />
     <Compile Include="UserControls\RevisionGrid\GitRefListsForRevisionTests.cs" />
     <Compile Include="UserControls\RevisionGrid\Graph\LaneNodeLocatorTests.cs" />
     <Compile Include="UserControls\RevisionGrid\Graph\LaneInfoProviderTests.cs" />
